+++ public/javascripts/scripts.js
@@ -1,3 +1,122 @@
+function setError(errid, succid, msg) {
+    if(!$("#" + succid).hasClass("hide")) {
+        $("#" + succid).addClass("hide");
+    }
+    $("#" + errid).removeClass("hide");
+    $("#" + errid).html(msg);
+}
+
+function setSuccess(errid, succid, msg) {
+    if(!$("#" + errid).hasClass("hide")) {
+        $("#" + errid).addClass("hide");
+    }
+    $("#" + succid).removeClass("hide");
+    $("#" + succid).html(msg);
+}
+
+function sendQuery(url, data, method, callback) {
+    $.ajax({
+        url: url,
+        dataType: "json",
+        contentType: "application/json; charset=utf-8",
+        data: JSON.stringify(data),
+        method: method
+    }).complete(function(xhr, textStatus) {
+        callback(xhr);
+    });
+}
+
+function createUser() {
+    var name = $("#name").val();
+    var password = $("#password").val();
+    var password2 = $("#passwordRepeat").val();
+    var level = $("#level").val();
+    if(name.length < 3) {
+        setError("createUserError", "createUserSuccess", "Name is too short. Should be at least 3 digits.");
+    }
+    else if($("#password").val() !== $("#passwordRepeat").val()) {
+        setError("createUserError","createUserSuccess", "Passwords do not match.");
+    }
+    else {
+        var cb = function(xhr) {
+            if(xhr.status == 200) {
+                updateManageUsers("createUserError", "createUserSuccess");
+                setTimeout(function() {setSuccess("createUserError","createUserSuccess", "User created.")}, 100);
+            }
+            else {
+                setError("createUserError","createUserSuccess", "Username already exists.");
+            }
+        };
+
+        var data = {name: name, password: password, level: level};
+        sendQuery("/account/create", data, "POST", cb);
+    }
+}
+
+function deleteUser(id, name) {
+    $("#popupTitle").text("Confirmation");
+    $("#popupBody").html("Are you sure you want to remove user " + name + "?");
+    $("#popupSave").html("Remove");
+    var cb = function(xhr) {
+        if(xhr.status) {
+            updateManageUsers("manageUsersError", "manageUsersSuccess");
+            setTimeout(function() {setSuccess("manageUsersError", "manageUsersSuccess", "User " + name + " successfully removed.")}, 100);
+        }
+        else {
+            setError("manageUsersError", "manageUsersSuccess", "Failed to remove " + name + ". Bad permissions.")
+        }
+        $("#popupSave").off('click');
+    };
+    $("#popupSave").on('click', function() {
+        sendQuery("/account/delete/" + id, null, "DELETE", cb);
+    });
+}
+
+function updateManageUsers(errid, succid) {
+    var cb = function(xhr) {
+        if(xhr.status == 200) {
+            $("#manageUsers").html(xhr.responseText);
+        }
+        else {
+            setError(errid, succid, "Failed to update manage users tab. Please reload manually.");
+        }
+    };
+    sendQuery("/account/users", null, "GET", cb);
+}
+
+function setAdmin(id, value) {
+    var cb = function(xhr) {
+        if(xhr.status === 200) {
+            if(parseInt(value) === 0) {
+                $("#" + id + "_admin").addClass("hide");
+                $("#" + id + "_notadmin").removeClass("hide");
+            }
+            else {
+                $("#" + id + "_notadmin").addClass("hide");
+                $("#" + id + "_admin").removeClass("hide");
+            }
+        }
+    };
+    sendQuery("/account/admin/" + id, {value: value}, "PUT", cb);
+}
+
+function setActive(id, value) {
+    var cb = function(xhr) {
+        if(xhr.status === 200) {
+            console.log("value = " + parseInt(value));
+            if(parseInt(value) === 0) {
+                $("#" + id + "_active").addClass("hide");
+                $("#" + id + "_inactive").removeClass("hide");
+            }
+            else {
+                $("#" + id + "_inactive").addClass("hide");
+                $("#" + id + "_active").removeClass("hide");
+            }
+        }
+    };
+    sendQuery("/account/active/" + id, {value: value}, "PUT", cb);
+}
+
 function updateAdminSettings() {
     var inputs = $("#adminSettings :input").not(":button");
     var data = [];
@@ -6,26 +125,19 @@ function updateAdminSettings() {
                     value: inputs[i].value
         });
     }
-
-    $.ajax({
-            url: "/account/adminsettings",
-            dataType: "json",
-            contentType: "application/json; charset=utf-8c",
-            data: JSON.stringify(data),
-            method: "POST"
-
-        }).complete(function(xhr, textStatus) {
-            if(xhr.status === 200) {
-                $("#error").html("User settings updated.");
-            }
-            else {
-                $("#error").html("Error updating user settings.");
-            }
-        });
+    var cb = function(xhr) {
+        if(xhr.status === 200) {
+            setSuccess("adminSettingsError", "adminSettingsSuccess", "Admin settings updated.");
+        }
+        else {
+            setError("adminSettingsError", "adminSettingsSuccess", "Error updating admin settings.");
+        }
+    };
+    sendQuery("/account/adminsettings", data, "POST", cb);
 }
 
 function updateUserPaths() {
-    var inputs = $("#userSettings :input").not(":button");
+    var inputs = $("#userPathsForm :checked").not(":button");
     var data = [];
     for(var i = 0; i < inputs.length; i++) {
         data.push({ id:  inputs[i].id,
@@ -33,47 +145,36 @@ function updateUserPaths() {
         });
     }
 
-    $.ajax({
-            url: "/account/usersettings",
-            dataType: "json",
-            contentType: "application/json; charset=utf-8c",
-            data: JSON.stringify(data),
-            method: "POST"
-
-        }).complete(function(xhr, textStatus) {
-            if(xhr.status === 200) {
-                $("#error").html("User settings updated.");
-            }
-            else {
-                $("#error").html("Error updating user settings.");
-            }
-        });
+    var cb = function(xhr) {
+        if(xhr.status === 200) {
+            setSuccess("userPathsError", "userPathsSuccess", "User paths updated.");
+        }
+        else {
+            setError("userPathsError", "userPathsSuccess", "Failed to update user paths.");
+        }
+    }
+    sendQuery("/account/userpaths", data, "POST", cb);
 }
 
 function changePassword() {
     if($("#newPassword").val() === $("#newPasswordConfirmation").val()) {
-        $.ajax({
-            url: "/account/settings",
-            dataType: "json",
-            contentType: "application/json; charset=utf-8c",
-            data: JSON.stringify({OldPassword: $("#oldPassword").val(), NewPassword: $("#newPassword").val()}),
-            method: "POST"
-
-        }).complete(function(xhr, textStatus) {
+        var cb = function(xhr) {
             if(xhr.status === 200) {
-                $("#error").html("Password CHANGED!!");
+                setSuccess("passwordError", "passwordSuccess", "Password successfully changed.");
             }
             else {
-                $("#error").html("BAD OLD PASSWORD");
+                setError("passwordError", "passwordSuccess", "Old password is incorrect.");
             }
-        });
+        };
+        var data = {OldPassword: $("#oldPassword").val(), NewPassword: $("#newPassword").val()};
+        sendQuery("/account/password", data, "POST", cb);
     }
     else {
-        $("#error").html("Passwords do not match");
+        setError("passwordError", "passwordSuccess", "Passwords do not match.");
     }
 }
 
-
+//TODO: REIMPLEMENT/FIX
 function initModalForRename(id, value) {
     $("#popupTitle").text("Name Change");
     $("#popupBody").html('<form>'
@@ -164,6 +265,7 @@ function initModalForMove(id) {
         
     });
 }
+
 $( document ).ready(function() {
     $('#settingsTabs a').click(function(e) {
         e.preventDefault();